# src/lib/openapi/__snapshots__/openapi-document.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PymtHouse Builder API",
    "version": "1.0.0",
    "description": "Machine-facing Builder API for integrator backends. OIDC issuer metadata (device flow, RFC 8693 token exchange at `/api/v1/oidc/token`) is published separately via OpenID Provider Configuration."
  },
  "components": {
    "schemas": {
...
      "OAuthError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "OAuth 2.0 error code (RFC 6749 §5.2).",
            "examples": [
              "invalid_request",
              "invalid_client",
              "invalid_scope"
            ]
          },
          "error_description": {
            "type": "string",
            "description": "Human-readable error description."
          },
          "correlation_id": {
            "type": "string",
            "format": "uuid",
            "description": "Request correlation id for support and audit."
          }
        },
        "required": [
          "error"
        ]
      },
...
      "SignerSession": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "description": "Short-lived signer JWT (RFC 8693)."
          },
          "token_type": {
            "type": "string",
            "enum": [
              "Bearer"
            ]
          },
          "expires_in": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true
          },
          "scope": {
            "type": "string",
            "minLength": 1,
            "description": "Space-delimited OAuth scopes.",
            "examples": [
              "sign:job"
            ]
          },
          "balanceUsdMicros": {
            "type": "string",
            "description": "Remaining balance in USD micros."
          },
          "lifetimeGrantedUsdMicros": {
            "type": "string",
            "description": "Lifetime granted balance in USD micros."
          },
          "signer_url": {
            "type": "string",
            "format": "uri",
            "description": "Public remote-signer base URL (snake_case canonical)."
          },
          "signerUrl": {
            "type": "string",
            "format": "uri",
            "description": "Deprecated alias of signer_url for legacy SDK consumers.",
            "deprecated": true
          },
          "issued_token_type": {
            "type": "string",
            "enum": [
              "urn:ietf:params:oauth:token-type:access_token"
            ],
            "description": "RFC 8693 issued_token_type when sourced from /token."
          },
          "correlation_id": {
            "type": "string",
            "format": "uuid",
            "description": "Request correlation id for support and audit."
          },
          "token": {
            "type": "object",
            "properties": {
              "accessToken": {
                "type": "string"
              },
              "access_token": {
                "type": "string"
              },
              "expiresIn": {
                "type": "number"
              },
              "expires_in": {
                "type": "number"
              },
              "scope": {
                "type": "string",
                "minLength": 1,
                "description": "Space-delimited OAuth scopes.",
                "examples": [
                  "sign:job"
                ]
              },
              "balanceUsdMicros": {
                "type": "string"
              },
              "lifetimeGrantedUsdMicros": {
                "type": "string"
              }
            },
            "required": [
              "access_token",
              "expires_in",
              "scope",
              "balanceUsdMicros",
              "lifetimeGrantedUsdMicros"
            ],
            "description": "Deprecated nested mirror; prefer top-level OAuth fields.",
            "deprecated": true
          }
        },
        "required": [
          "access_token",
          "token_type",
          "expires_in",
          "scope",
          "balanceUsdMicros",
          "lifetimeGrantedUsdMicros"
        ]
      },
      "ApiKeySignerSessionRequest": {
        "type": "object",
        "properties": {
          "scope": {
            "type": "string",
            "minLength": 1,
            "description": "Space-delimited OAuth scopes.",
            "examples": [
              "sign:job"
            ]
          }
        }
      },
...
      "LegacyValidateResponse": {
        "type": "object",
        "properties": {
          "valid": {
            "type": "boolean"
          },
          "client_id": {
            "type": "string"
          },
          "plan": {
            "nullable": true
          },
          "allowedModels": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "subscriptionRef": {
            "type": "string"
          }
        },
        "required": [
          "valid"
        ]
      },
      "C0ValidateResponse": {
        "type": "object",
        "properties": {
          "valid": {
            "type": "boolean"
          },
          "user": {
            "type": "object",
            "properties": {
              "sub": {
                "type": "string"
              }
            },
            "required": [
              "sub"
            ]
          },
          "billing_account": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "providerSlug": {
                "type": "string"
              },
              "billingMode": {
                "type": "string",
                "enum": [
                  "delegated",
                  "prepay"
                ]
              }
            },
            "required": [
              "id",
              "providerSlug",
              "billingMode"
            ]
          },
          "capabilities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "quota": {
            "nullable": true
          },
          "subscriptionRef": {
            "type": "string"
          }
        },
        "required": [
          "valid"
        ]
      },
      "C0ValidateRequest": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "key"
        ]
      },
      "AdminTokenIssueResponse": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string"
          },
          "token": {
            "type": "string"
          },
          "scopes": {
            "type": "string"
          },
          "endUserId": {
            "type": "string",
            "nullable": true
          },
          "expiresInDays": {
            "type": "number"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "sessionId",
          "token",
          "scopes",
          "endUserId",
          "expiresInDays",
          "message"
        ]
      },
      "AdminTokenIssueRequest": {
        "type": "object",
        "properties": {
          "scopes": {
            "type": "string"
          },
          "expiresInDays": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true
          },
          "endUserId": {
            "type": "string"
          },
          "label": {
            "type": "string"
          }
        }
      },
      "AdminTokenListResponse": {
        "type": "object",
        "properties": {
          "tokens": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "label": {
                  "type": "string",
                  "nullable": true
                },
                "endUserId": {
                  "type": "string",
                  "nullable": true
                },
                "scopes": {
                  "type": "string"
                },
                "expiresAt": {
                  "type": "string",
                  "nullable": true
                },
                "createdAt": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "label",
                "endUserId",
                "scopes",
                "expiresAt",
                "createdAt"
              ]
            }
          }
        },
        "required": [
          "tokens"
        ]
      },
      "AdminTokenRevokeRequest": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "sessionId"
        ]
      }
    },
...
    "securitySchemes": {
      "m2mBasic": {
        "type": "http",
        "scheme": "basic",
        "description": "Confidential M2M client (`m2m_…` + `pmth_cs_…` secret). RFC 6749 client authentication."
      },
      "bearerApiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Per-app-user API key (`pmth_<hex>`). Not an M2M client secret."
      },
      "bearerUserJwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Short-lived user access token minted by Builder API or OIDC."
      },
      "adminSession": {
        "type": "http",
        "scheme": "bearer",
        "description": "Dashboard admin session or admin-scoped platform token."
      }
    }
  },
  "paths": {
...
    "/api/v1/apps/{clientId}/auth/api-key/signer-session": {
      "post": {
        "tags": [
          "Credentials"
        ],
        "summary": "Exchange API key for signer session (canonical)",
        "description": "Single-call path: `pmth_*` API key → signer JWT via internal user-token mint and RFC 8693-equivalent exchange. Preferred over dashboard BFF `/api/pymthouse/keys/exchange`.",
        "security": [
          {
            "bearerApiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^app_[a-f0-9]+$",
              "description": "Public Builder app client id (`app_…`).",
              "examples": [
                "app_3b386c81a1db1169fd2c3986"
              ]
            },
            "required": true,
            "description": "Public Builder app client id (`app_…`).",
            "name": "clientId",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiKeySignerSessionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Canonical SignerSession envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignerSession"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthError"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthError"
                }
              }
            }
          },
          "404": {
            "description": "App not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/validate": {
      "get": {
        "tags": [
          "Credentials"
        ],
        "summary": "Validate API key (legacy GET)",
        "deprecated": true,
        "description": "Deprecated: Bearer validation for legacy BPP consumers. Prefer POST validate (C0) when enabled.",
        "security": [
          {
            "bearerApiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Validation result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyValidateResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid key"
          }
        }
      },
      "post": {
        "tags": [
          "Credentials"
        ],
        "summary": "Validate API key (C0 POST)",
        "description": "Provider-neutral validate when `BPP_VALIDATE_V2=1`.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/C0ValidateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "C0 validate body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/C0ValidateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing key"
          },
          "401": {
            "description": "Invalid key"
          },
          "404": {
            "description": "Feature disabled"
          }
        }
      }
    },
    "/api/v1/tokens": {
      "post": {
        "tags": [
          "Platform"
        ],
        "summary": "Issue admin/platform bearer token",
        "security": [
          {
            "adminSession": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminTokenIssueRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Issued token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminTokenIssueResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "tags": [
          "Platform"
        ],
        "summary": "List admin/platform tokens",
        "security": [
          {
            "adminSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Token list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminTokenListResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Platform"
        ],
        "summary": "Revoke admin/platform token",
        "security": [
          {
            "adminSession": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminTokenRevokeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Revoked"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "PymtHouse API origin"
    }
  ],
...
}